Refactorización: 'enums' de HabilityResources.h y RPGLogic.h a 'enum class' (+ sprites de combate)
Además:
1. Cambio de sintaxis de algunas palabras claves.
2. Todo lo que conlleva y los cambios en el código de acceso a las variables

Por otro lado:
1. El enemigo se muestra con su vida (mejorable)
2. Se procesan las eliminaciones correctamente
3. Las habilidades se muestran por su imagen correspondiente
4. Los targets ya sean enemigos o aliados se muestran correctamente

+++ Proyecto.02/project/src/Templates/CharacterSheet.cpp
@@ -42,7 +42,7 @@ void CharacterSheet::recieveMana(int mana)
 {
 	//cambiar el _hitPoints por el stat que aumenta
 
-	_stats[st].value += buff;
+	_stats[size_t(st)].value += buff;
 
 	//cout << name << " aumenta su "<< stat << buff << " puntos" << endl;
